# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing Pointlings, users, items, and XP system",
        "title": "Pointlings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pointlings/{pointling_id}": {
            "get": {
                "description": "Get detailed information about a specific pointling",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pointlings"
                ],
                "summary": "Get pointling details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pointling"
                        }
                    },
                    "400": {
                        "description": "Invalid pointling ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pointling not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pointlings/{pointling_id}/xp": {
            "post": {
                "description": "Award experience points to a pointling from a specific source",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "XP"
                ],
                "summary": "Add XP to pointling",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "XP addition request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.addXPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.addXPResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid pointling ID, request body, or XP amount",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pointling not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Daily XP limit reached for this source",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pointlings/{pointling_id}/xp/history": {
            "get": {
                "description": "Get paginated history of XP gains for a pointling",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "XP"
                ],
                "summary": "Get XP history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of records to return (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.XPEvent"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid pointling ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handlers.addXPRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 100
                },
                "source": {
                    "enum": [
                        "QUEST_COMPLETE",
                        "DAILY_LOGIN",
                        "ACHIEVEMENT"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.XPEventSource"
                        }
                    ],
                    "example": "QUEST_COMPLETE"
                }
            }
        },
        "handlers.addXPResponse": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/models.XPEvent"
                },
                "leveled_up": {
                    "type": "boolean",
                    "example": true
                },
                "new_level": {
                    "type": "integer",
                    "example": 5
                },
                "new_total": {
                    "type": "integer",
                    "example": 1250
                },
                "pointling_id": {
                    "type": "integer",
                    "example": 123
                },
                "required_xp": {
                    "type": "integer",
                    "example": 2000
                },
                "xp_gained": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.JSONMap": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.Pointling": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_xp": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "look_json": {
                    "$ref": "#/definitions/models.JSONMap"
                },
                "nickname": {
                    "type": "string"
                },
                "personality_id": {
                    "type": "integer"
                },
                "pointling_id": {
                    "type": "integer"
                },
                "required_xp": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.XPEvent": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "event_ts": {
                    "type": "string"
                },
                "pointling_id": {
                    "type": "integer"
                },
                "source": {
                    "$ref": "#/definitions/models.XPEventSource"
                },
                "xp_amount": {
                    "type": "integer"
                }
            }
        },
        "models.XPEventSource": {
            "type": "string",
            "enum": [
                "RECEIPT",
                "PLAY",
                "DAILY"
            ],
            "x-enum-varnames": [
                "XPSourceReceipt",
                "XPSourcePlay",
                "XPSourceDaily"
            ]
        }
    },
...
}